Added CmeMessageBlockWriter, which writes MsgSeqNum into the first four bytes of the CME preamble.  Currently, it always writes a Channel Sub ID (the fifth byte) of zero.
A slight change to the writeBlockLength message signature was required to
accommodate the CmeMessageBlockWriter.

The CmeMessageBlockReader will now actually read the CME preamble instead of
just skipping it.

The producer and consumer examples now take a 'variant' argument (-z) which
causes them to enable exchange-specific behavior.  Currently, '-z CME' is the
only supported variant.


git-svn-id: https://svn.code.sf.net/p/openfast/code/trunk@230 e6ee02d6-d133-0410-bd08-f9337e8129a4